pyproject.toml
[tool.poetry.scripts]
main = "rp_utilities.__main__:main"

[tool.poetry]
name = "rp-utilities"
version = "0.1.0"
description = ""
authors = ["HirukaRogue <140905381+HirukaRogue@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "rp_utilities"}]

[tool.poetry.dependencies]
python = "^3.11"
"discord.py" = {git = "https://github.com/Rapptz/discord.py"}
"discord-ext-menus" = {git = "https://github.com/Rapptz/discord-ext-menus"}
python-dotenv = "^1.0.0"
sympy = "^1.12"
jishaku = "^2.5.2"
pyright = "^1.1.347"
lark = "^1.1.9"
aiosqlite = "^0.20.0"
aiopyston = "^1.2.1"
requests = "^2.32.3"


[tool.poetry.group.dev.dependencies]
black = "^23.12.1"
ruff = "^0.1.12"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100

[tool.mypy]
python_version = 3.11
strict = true
pretty = true
install_types = true
ignore_missing_imports = true
ignore_missing_imports_per_module = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_calls = true
disallow_untyped_decorators = true
strict_equality = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
warn_unreachable = true
show_error_context = true
show_column_numbers = true
show_error_codes = true

[tool.black]
line-length = 100
target-version = ["py311"]